🤖 Refactor(client): dimension Context 참조 에러 해결
