+++ apps/client/src/components/Layout/Sidebar/SelectPlatform.tsx
