+++ apps/client/src/components/Node/ncloud/ObjectStorageNode.tsx
@@ -1,4 +1,5 @@
-import { useCanvasDimensionContext } from '@contexts/CanvasDimensionContext';
+import { useDimensionContext } from '@contexts/DimensionContext';
+import { Node } from '@types';
 
 const Node3D = () => {
     return (
@@ -56,7 +57,7 @@ const Node2D = () => {
         </svg>
     );
 };
-export default () => {
-    const { dimension } = useCanvasDimensionContext();
+export default ({}: Partial<Node>) => {
+    const { dimension } = useDimensionContext();
     return dimension === '2d' ? <Node2D /> : <Node3D />;
 };
