+++ apps/client/src/components/Node/ncloud/ServerNode.tsx
@@ -1,4 +1,7 @@
-import { useCanvasDimensionContext } from '@contexts/CanvasDimensionContext';
+import { useDimensionContext } from '@contexts/DimensionContext';
+import { Node } from '@types';
+
+type Props = {};
 
 const Node3D = () => {
     return (
@@ -87,7 +90,7 @@ const Node2D = () => {
     );
 };
 
-export default () => {
-    const { dimension } = useCanvasDimensionContext();
+export default ({}: Partial<Node>) => {
+    const { dimension } = useDimensionContext();
     return dimension === '2d' ? <Node2D /> : <Node3D />;
 };
